requirements.txt
Django==4.2.7
Pillow==10.0.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-ckeditor==6.7.0
django-taggit==5.0.0
django-allauth==0.58.2
django-debug-toolbar==4.2.0
python-dotenv==1.0.0